+++ src/navigation/AppNavigator.tsx
@@ -21,7 +21,7 @@ import {MyAccountScreen} from 'screen/MyAccountScreen';
 import MyIdentityScreen from 'screen/MyIdentityScreen';
 import {IdentityGuideScreen} from 'screen/MyIdentityScreen/IdentityGuideScreen';
 import {NotificationSettingsScreen} from 'screen/NotificationSettingsScreen';
-import {CrowdLoanScreen} from 'screen/Parachain/CrowdLoanScreen';
+import {CrowdLoanScreen} from 'screen/Parachains/CrowdLoanScreen';
 import {PermissionGrantingPrompt} from 'screen/PermissionGrantingPrompt';
 import {PolkassemblyDiscussionDetail} from 'screen/Polkassembly/PolkassemblyDiscussionDetail';
 import {PolkassemblyDiscussions} from 'screen/Polkassembly/PolkassemblyDiscussions';
@@ -48,7 +48,7 @@ import {
 import * as routeKeys from 'src/navigation/routeKeys';
 import {darkTheme, lightTheme} from 'src/navigation/theme';
 import globalStyles from 'src/styles';
-import {CrowdLoanFundDetailScreen as CrowdLoanFundDetailScreen} from 'screen/Parachain/CrowdLoanFundDetailScreen';
+import {CrowdLoanFundDetailScreen as CrowdLoanFundDetailScreen} from 'screen/Parachains/CrowdLoanFundDetailScreen';
 
 const DashboardStack = createStackNavigator<DashboardStackParamList>();
